Reordered InventoryFormat, no need to have different formats now
